deepscanscape: Check and Annotate Deep Mutational Scanning Data

The deepscanscape package provides functions to standardise deep mutational scanning data and compare it to a deep mutational landscape calculated from a dataset of 28 previous studies. This allows the whole dataset to be checked for unusual properties and individual positions to be annotated with positional subtypes, indicating the positions in other proteins they are most similar to.

Data Processing Functions

deep_mutational_scan - Construct standardised deep mutational scan datasets. This creates an S3 object with various generics.

check_data - Check a deep_mutational_scan for common abnormalities

bind_scans - Combined deep mutational scan datasets

parse_deep_scan - Parse common deep scan data formats

transform_er - Transform ER scores from common score types to a standardised scale

normalise_er - Normalise ER scores

impute - Impute missing data from deep mutational scans

Annotation Functions

annotate - Add annotations from the combined landscape to deep mutational scan data

describe_clusters - Add details on positions assigned clusters

landscape_outliers - Identify rows of a deep_mutational_scan that lie away from the studied regions of the deep landscape.

recluster - Perform the original clustering procedure on a new deep mutational scan dataset

Visualisation Functions

plot_er_distribution - Compare the distribution of ER scores in new data to the deep landscape dataset.

plot_er_heatmap - Plot heatmaps show fitness scores across a protein

plot_landscape - Project a new dataset onto the deep mutational landscape, including visualising various biophysical properties.

plot_cluster_frequencies - Plot the frequencies of amino acid subtypes in a new dataset

plot_recluster - Summarise the profiles of a new clustered dataset