Describe amino acid positional subtypes

Add descriptions and frequencies from the cluster assigned to each position in an annotated deep_mutational_scan dataset. These were determined through analysis of the larger deep_landscape dataset. Numerical summary statistics based on the same dataset can also be added, which give the mean characteristics of the subtype. They include results from FoldX and Naccess and mean ER fitness score profiles.

describe_clusters(x, full = FALSE)

Arguments

x
deep_mutational_scan.

full
Logical. Include average statistics from the deep_landscape dataset in addition to summary descriptions and notes.

Value

A tibble, with each row detailing a row of the input data and columns matching those in the subtypes dataset.

Details

The deep landscape properties summarised where chosen by benchmarking subtype based predictions against the landscape data and a separate deep mutational scan of SARS-CoV-2 Spike by Starr et al. (2020). Metrics are only shown when the average value of positions of that subtype is meaningfully related to the observed values.

Properties are not shown for outlier subtypes, which do not have consistent properties, and ambiguously assigned positions, since the properties could be that of either of the subtypes the position could have been assigned to.

Examples

dms <- deepscanscape::deep_scans$p53

# Basic description
describe_clusters(dms)

#> Warning: deep_mutational_scan is not annotated. Annotating using annotate_dms().

#> # A tibble: 191 x 8
#>    name   position wt    cluster global_cluster_f… group description    notes   
#>    <chr>     <dbl> <chr> <chr>               <dbl> <chr> <chr>          <chr>   
#>  1 Kotle…        1 T     T5                0.0615  Nega… Weak selectio… Weak se…
#>  2 Kotle…        2 Y     Y3                0.227   Arom… Generally sel… General…
#>  3 Kotle…        3 Q     Q2                0.2     Not … Anything but … Anythin…
#>  4 Kotle…        4 G     G6                0.0512  Uniq… Largely permi… Largely…
#>  5 Kotle…        5 S     S1                0.398   Smal… Only tolerate… Only to…
#>  6 Kotle…        6 Y     Y3                0.227   Arom… Generally sel… General…
#>  7 Kotle…        7 G     G6                0.0512  Uniq… Largely permi… Largely…
#>  8 Kotle…        8 F     FO                0.00722 Outl… Outlier        NA      
#>  9 Kotle…        9 R     R1                0.461   Posi… Intolerant of… Intoler…
#> 10 Kotle…       10 L     LO                0.0561  Outl… Outlier        NA      
#> # … with 181 more rows

# More details
describe_clusters(dms, full = TRUE)

#> Warning: deep_mutational_scan is not annotated. Annotating using annotate_dms().

#> # A tibble: 191 x 46
#>    name  position wt    cluster global_cluster_… group description notes       A
#>    <chr>    <dbl> <chr> <chr>              <dbl> <chr> <chr>       <chr>   <dbl>
#>  1 Kotl…        1 T     T5               0.0615  Nega… Weak selec… Weak… -0.0596
#>  2 Kotl…        2 Y     Y3               0.227   Arom… Generally … Gene… -0.550 
#>  3 Kotl…        3 Q     Q2               0.2     Not … Anything b… Anyt… -0.0168
#>  4 Kotl…        4 G     G6               0.0512  Uniq… Largely pe… Larg…  0.0407
#>  5 Kotl…        5 S     S1               0.398   Smal… Only toler… Only… -0.247 
#>  6 Kotl…        6 Y     Y3               0.227   Arom… Generally … Gene… -0.550 
#>  7 Kotl…        7 G     G6               0.0512  Uniq… Largely pe… Larg…  0.0407
#>  8 Kotl…        8 F     FO               0.00722 Outl… Outlier     NA    NA     
#>  9 Kotl…        9 R     R1               0.461   Posi… Intolerant… Into… -0.605 
#> 10 Kotl…       10 L     LO               0.0561  Outl… Outlier     NA    NA     
#> # … with 181 more rows, and 37 more variables: C <dbl>, D <dbl>, E <dbl>,
#> #   F <dbl>, G <dbl>, H <dbl>, I <dbl>, K <dbl>, L <dbl>, M <dbl>, N <dbl>,
#> #   P <dbl>, Q <dbl>, R <dbl>, S <dbl>, T <dbl>, V <dbl>, W <dbl>, Y <dbl>,
#> #   mean_score <dbl>, total_energy <dbl>, backbone_hbond <dbl>,
#> #   sidechain_hbond <dbl>, van_der_waals <dbl>, electrostatics <dbl>,
#> #   solvation_polar <dbl>, solvation_hydrophobic <dbl>,
#> #   van_der_waals_clashes <dbl>, entropy_sidechain <dbl>,
#> #   entropy_mainchain <dbl>, cis_bond <dbl>, torsional_clash <dbl>,
#> #   backbone_clash <dbl>, disulfide <dbl>, partial_covalent_bonds <dbl>,
#> #   energy_ionisation <dbl>, all_atom_rel <dbl>