Generate an HTML report for new deep mutational scan data

Perform a full analysis on new deep mutational scan data and summarise the results as an HTML report. The report provides an overview of the new data and then compares the properties of the new data to the datasets included in the deep_landscape dataset. This helps to identify unusual properties, map the features of the protein on the common mutational landscape and assign each position to an amino acid subtype, which informs on its properties.

generate_report(
  ...,
  output_path = "deepscanscape_report.html",
  highlight = NULL
)

Arguments

...
deep_landscape objects or lists of such objects to report on.

output_path
Path to write report to.

highlight
A named list of positions to highlight, where each element is named for a study in ... and contains the positions in that study to highlight.

Value

Path to the generated report.

Examples

if (FALSE) {
  # For a single study
  generate_report(deep_scans$p53, highlight = c(50, 100, 110))

  # For multiple studies
  highlight <- list(`Elazar GpA`=c(1,2,3), `Kotler p53`=c(50, 100, 110))
  generate_report(deep_scans, highlight = highlight)
}