Amino acid positional subtypes from the deep mutational landscape

A dataset containing groups, descriptions and notes about each amino acid subtype identified in the deep mutational landscape. Groups were determined by ER profile correlation and descriptions/notes from manual assessment of subtype position properties (ER, FoldX predicitons, conservation, etc.).

subtypes

Format

A data frame with 119 rows and 5 columns

wt

Wild type amino acid

cluster

Amino acid subtype

prop

Proportion of amino acid positions in the original dataset that are classified into this cluster

group

Subtype group

description

Description of the subtype

notes

Additional notes about positions of this subtype

A

Normalised fitness score when mutating to alanine

C

Normalised fitness score when mutating to cysteine

D

Normalised fitness score when mutating to aspartate

E

Normalised fitness score when mutating to glutamate

F

Normalised fitness score when mutating to phenylalanine

G

Normalised fitness score when mutating to glycine

H

Normalised fitness score when mutating to histidine

I

Normalised fitness score when mutating to isoleucine

K

Normalised fitness score when mutating to lysine

L

Normalised fitness score when mutating to leucine

M

Normalised fitness score when mutating to methionine

N

Normalised fitness score when mutating to asparagine

P

Normalised fitness score when mutating to proline

Q

Normalised fitness score when mutating to glutamine

R

Normalised fitness score when mutating to arginine

S

Normalised fitness score when mutating to serine

T

Normalised fitness score when mutating to threonine

V

Normalised fitness score when mutating to valine

W

Normalised fitness score when mutating to tryptophan

Y

Normalised fitness score when mutating to tyrosine

mean_score

Mean normalised ER fitness score

mean_sift

Mean SIFT4G score for positions of this cluster

total_energy

Mean FoldX total ddG for mutations to position of this cluster

backbone_hbond

Mean FoldX backbone_hbond term

sidechain_hbond

Mean FoldX sidechain_hbond term

van_der_waals

Mean FoldX van_der_waals term

electrostatics

Mean FoldX electrostatics term

solvation_polar

Mean FoldX solvation_polar term

solvation_hydrophobic

Mean FoldX solvation_hydrophobic term

van_der_waals_clashes

Mean FoldX van_der_waals_clashes term

entropy_sidechain

Mean FoldX entropy_sidechain term

entropy_mainchain

Mean FoldX entropy_mainchain term

cis_bond

Mean FoldX cis_bond term

torsional_clash

Mean FoldX torsional_clash term

backbone_clash

Mean FoldX backbone_clash term

helix_dipole

Mean FoldX helix_dipole term

disulfide

Mean FoldX disulfide term

electrostatic_kon

Mean FoldX electrostatic_kon term

partial_covalent_bonds

Mean FoldX partial_covalent_bonds term

energy_ionisation

Mean FoldX energy_ionisation term

all_atom_rel

Mean relative surface accessibility

Source

Dunham and Beltrao (2020)